LayoutTests/editing/execCommand/insert-list-nested-with-orphaned-live-range.html

<script src="../../resources/dump-as-markup.js"></script><!-- webkit-test-runner [ LiveRangeSelectionEnabled=true ] -->
<div id="test" contenteditable>
<ul>
    <ul><li>hello</li></ul>
    world
    <ol><li>WebKit</li></ol>
    <ul>
        <li>rocks</li>
        <ul><li>because of you</li></ul>
    </ul>
</ul>
</div>
<script type="text/javascript"> 

if (window.testRunner)
    testRunner.dumpAsText();

window.getSelection().selectAllChildren(document.getElementById('test'))
document.execCommand('InsertOrderedList', false, null);

Markup.description("This tests hang when listifying nested lists with an orphaned list child in between (see bug 33668).")
Markup.dump('test')
</script>
